web/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["rtatton <rtatton17@proton.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.3"
hat-py-sdk = { git = "https://github.com/rtatton/hat-py-sdk", rev = "v0.12.1" }
keyring = "^23.7.0"
pydantic = "^1.9.1"
Flask-Security-Too = "^4.1.5"
python-dotenv = "^0.20.0"
email-validator = "^1.2.1"
redis-om = "^0.0.27"
zxcvbn = "^4.4.28"
bleach = "^5.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
